add ids for buttons
