+++ f2/src/forms/ContactInfoForm.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { jsx } from '@emotion/core'
 import { Trans } from '@lingui/macro'
 import { Form, Field } from 'react-final-form'
-import { NextAndCancelButtons } from '../components/buttons-container'
+import { NextAndCancelButtons } from '../components/next-and-cancel-buttons'
 import { TextInput } from '../components/TextInput'
 import { Text } from '../components/text'
 import { ApolloConsumer } from 'react-apollo'
@@ -72,9 +72,7 @@ export const ContactInfoForm = ({ onSubmit }) => (
             </div>
 
             <NextAndCancelButtons>
-              <Trans id="contactinfoPage.nextButton">
-                Next: Confirm information
-              </Trans>
+              <Trans id="contactinfoPage.nextButton" />
             </NextAndCancelButtons>
           </form>
         )}
