+++ f2/src/forms/ImpactStatementInfoForm.js
@@ -5,7 +5,7 @@ import { ApolloConsumer } from 'react-apollo'
 import { useLingui } from '@lingui/react'
 import { Trans } from '@lingui/macro'
 import { Form, Field } from 'react-final-form'
-import { NextAndCancelButtons } from '../components/buttons-container'
+import { NextAndCancelButtons } from '../components/next-and-cancel-buttons'
 import { Checkbox } from '../components/checkbox'
 import { Text } from '../components/text'
 import { TextArea } from '../components/text-area'
