+++ f2/src/locales/en.json
@@ -122,6 +122,7 @@
   "impactPage.summary": "Tell us more about how it impacted you.",
   "impactPage.title": "Impact caused by the scam",
   "landingPage.intro": "Scams can happen to anyone. They’re not always easy to recognize and new ones are invented every day.",
+  "landingPage.nextButton": "Report now ❯",
   "landingPage.required0": "Before you start, gather what you can about the scam:",
   "landingPage.required1": "Dates when it took place",
   "landingPage.required2": "Descriptions of what happened",
