+++ f2/src/locales/fr.json
@@ -122,6 +122,7 @@
   "impactPage.summary": "Dites-nous en plus sur l'impact que cela a eu sur vous.",
   "impactPage.title": "Impact causé par l'escroquerie",
   "landingPage.intro": "N'importe qui peut être victime d'une fraude. De nouvelles fraudes sont inventées chaque jour, et elles ne sont pas toujours facile à reconnaître.",
+  "landingPage.nextButton": "Signaler maintenant ❯",
   "landingPage.required0": "Avant de commencer, recueillez les informations que vous pouvez au sujet de la fraude:",
   "landingPage.required1": "Dates auxquelles la fraude a eu lieu",
   "landingPage.required2": "Descriptions de ce qui s'est passé",
